+++ R/class-workbook-utils.R
@@ -148,11 +148,45 @@ validate_cf_params <- function(params) {
   params
 }
 
+#' create required columns
+#' @param wb a workbook object
+#' @param sheet a worksheet
+#' @param cols the required columns
+#' @noRd
+wb_create_columns <- function(wb, sheet, cols) {
+
+  col_df <- wb$worksheets[[sheet]]$unfold_cols()
+
+  # create empty cols
+  if (NROW(col_df) == 0)
+    col_df <- col_to_df(read_xml(wb$createCols(sheet, n = max(cols))))
+
+  # found a few cols, but not all required cols. create the missing columns
+  if (any(!cols %in% as.numeric(col_df$min))) {
+    beg <- max(as.numeric(col_df$min)) + 1
+    end <- max(cols)
+
+    # new columns
+    new_cols <- col_to_df(read_xml(wb$createCols(sheet, beg = beg, end = end)))
+
+    # rbind only the missing columns. avoiding dups
+    sel <- !new_cols$min %in% col_df$min
+    col_df <- rbind(col_df, new_cols[sel, ])
+    col_df <- col_df[order(as.numeric(col_df[, "min"])), ]
+  }
+
+  col_df
+}
+
 # waivers -----------------------------------------------------------------
 
 #' `openxlsx2` waivers
 #'
-#' Waiver functions for `openxlsx2` functions
+#' Waiver functions for `openxlsx2` functions.
+#' * `current_sheet()` uses [wb_get_active_sheet()] by default if performing
+#'   actions on a worksheet, for example when you add data.
+#' * `next_sheet()` is used when you add a new worksheet, a new chartsheet or
+#'   when you add a pivot table
 #'
 #' @name waivers
 #' @returns An object of class `openxlsx2_waiver`